+++ mobile/Screens/SignUpScreen.jsx
@@ -127,7 +127,7 @@ export default function SignUp({ navigation }) {
         console.error(res.data.error);
       } else {
         const userId = res.data._id;
-        Storage({ key: 'hello', value: userId, saveKey: true });
+        await Storage({ key: 'userId', value: userId, saveKey: true });
         navigation.navigate('Home');
       }
     } catch (err) {
